requirements.txt
diffusers==0.23.0
einops==0.3.0
fvcore==0.1.5.post20221221
gradio==3.39.0
huggingface_hub==0.20.1
numpy==1.23.1
omegaconf==2.1.1
opencv_contrib_python
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
packaging==24.0
Pillow==10.0.1
safetensors==0.4.3
setuptools==68.2.2
submitit==1.5.1
torch==2.0.1
torchmetrics==0.6.0
torchvision==0.15.2
transformers==4.25.1
accelerate==0.25.0
modelscope

